Move FlatMap, plot_swanson, plot_swanson_vector
