+++ ibllib/atlas/atlas.py
@@ -1549,102 +1549,6 @@ def _download_atlas_allen(target_file_image):
     return Path(http_download_file(url, target_dir=target_file_image.parent))
 
 
-class FlatMap(AllenAtlas):
-    """The Allen Atlas flatmap."""
-
-    def __init__(self, flatmap='dorsal_cortex', res_um=25):
-        """
-        Avaiable flatmaps are currently 'dorsal_cortex', 'circles' and 'pyramid'
-        :param flatmap:
-        :param res_um:
-        """
-        super().__init__(res_um=res_um)
-        self.name = flatmap
-        if flatmap == 'dorsal_cortex':
-            self._get_flatmap_from_file()
-        elif flatmap == 'circles':
-            from ibllib.atlas.flatmaps import circles
-            if res_um != 25:
-                raise NotImplementedError('Pyramid circles not implemented for resolution other than 25um')
-            self.flatmap, self.ml_scale, self.ap_scale = circles(N=5, atlas=self, display='flat')
-        elif flatmap == 'pyramid':
-            from ibllib.atlas.flatmaps import circles
-            if res_um != 25:
-                raise NotImplementedError('Pyramid circles not implemented for resolution other than 25um')
-            self.flatmap, self.ml_scale, self.ap_scale = circles(N=5, atlas=self, display='pyramid')
-
-    def _get_flatmap_from_file(self):
-        # gets the file in the ONE cache for the flatmap name in the property, downloads it if needed
-        file_flatmap = self._get_cache_dir().joinpath(f'{self.name}_{self.res_um}.nrrd')
-        if not file_flatmap.exists():
-            file_flatmap.parent.mkdir(exist_ok=True, parents=True)
-            aws.s3_download_file(f'atlas/{file_flatmap.name}', file_flatmap)
-        self.flatmap, _ = nrrd.read(file_flatmap)
-
-    def plot_flatmap(self, depth=0, volume='annotation', mapping='Allen', region_values=None, ax=None, **kwargs):
-        """
-        Displays the 2D image corresponding to the flatmap.
-
-        If there are several depths, by default it will display the first one.
-
-        Parameters
-        ----------
-        depth : int
-            Index of the depth to display in the flatmap volume (the last dimension).
-        volume : {'image', 'annotation', 'boundary', 'value'}
-            - 'image' - Allen image volume.
-            - 'annotation' - Allen annotation volume.
-            - 'boundary' - outline of boundaries between all regions.
-            - 'volume' - custom volume, must pass in volume of shape BrainAtlas.image.shape as
-               regions_value argument.
-        mapping : str, default='Allen'
-            The brain region mapping to use.
-        region_values : numpy.array
-            An array the shape of the brain atlas image containing custom region values. Used when
-            `volume` value is 'volume'.
-        ax : matplotlib.pyplot.Axes, optional
-            A set of axes to plot to.
-        **kwargs
-            See matplotlib.pyplot.imshow.
-
-        Returns
-        -------
-        matplotlib.pyplot.Axes
-            The plotted image axes.
-        """
-        if self.flatmap.ndim == 3:
-            inds = np.int32(self.flatmap[:, :, depth])
-        else:
-            inds = np.int32(self.flatmap[:, :])
-        regions = self._get_mapping(mapping=mapping)[self.label.flat[inds]]
-        if volume == 'annotation':
-            im = self._label2rgb(regions)
-        elif volume == 'value':
-            im = region_values[regions]
-        elif volume == 'boundary':
-            im = self.compute_boundaries(regions)
-        elif volume == 'image':
-            im = self.image.flat[inds]
-        else:
-            raise ValueError(f'Volume type "{volume}" not supported')
-        if not ax:
-            ax = plt.gca()
-
-        return self._plot_slice(im, self.extent_flmap(), ax=ax, volume=volume, **kwargs)
-
-    def extent_flmap(self):
-        """
-        Returns the boundary coordinates of the flat map.
-
-        Returns
-        -------
-        numpy.array
-            The bounding coordinates of the flat map image, specified as (left, right, bottom, top).
-        """
-        extent = np.r_[0, self.flatmap.shape[1], 0, self.flatmap.shape[0]]
-        return extent
-
-
 class FranklinPaxinosAtlas(BrainAtlas):
     """
     The Franklin & Paxinos brain atlas.
